Reading MSP files The readMsp function imports the data from a file in MGF format reading all specified fields and returning the data as a DataFrame().

Reading MSP files

The readMsp function imports the data from a file in MGF format reading all specified fields and returning the data as a DataFrame().

readMsp(f, msLevel = 2L, mapping = spectraVariableMapping(MsBackendMsp()), ...)

Arguments

f

character(1) with the path to an MSP file.

msLevel

numeric(1) with the MS level. Default is 2. This value will be reported as the spectra's MS level unless the source MSP file defines the MS level.

mapping

named character vector to rename MSP fields to spectra variables (see spectraVariableMapping()). This allows to correctly import also custom fields or data from files with different MSP flavors.

...

Additional parameters, currently ignored.

Value

A DataFrame with each row containing the data from one spectrum in the MSP file. m/z and intensity values are available in columns "mz"and "intensity" in a list representation.

Author

Laurent Gatto, Steffen Neumann, Johannes Rainer

Examples

fls <- dir(system.file("extdata", package = "MsBackendMsp"),
    full.names = TRUE, pattern = "msp$")[1L]

readMsp(fls)
#> DataFrame with 1 row and 20 columns
#>   Collision_energy               Comments         ExactMass      Formula
#>        <character>            <character>       <character>  <character>
#> 1             45 V "SMILES=CCCC(=O)OC(C.. 232.1543346040907 [C11H22NO4]+
#>                 InChIKey  Instrument Instrument_type    Ion_mode          MW
#>              <character> <character>     <character> <character> <character>
#> 1 QWYFHHGCZUCMBN-UHFFF..  SCIEX 5600  in-silico QTOF           P         232
#>          Name   Num.Peaks PrecursorMZ Precursor_type Spectrum_type
#>   <character> <character> <character>    <character>   <character>
#> 1    ACar 4:0           2   232.15433           [M]+           MS2
#>          accession              Synon                mz         intensity
#>        <character>             <list>     <NumericList>     <NumericList>
#> 1 LipidBlast000001 [M]+,$:00in-source  85.0284,232.1543  80.0801,100.0000
#>               dataOrigin   msLevel
#>              <character> <integer>
#> 1 /__w/_temp/Library/M..         2